# account-service/openapi.yaml
openapi: 3.0.1
info:
  title: "Account Service - TraderX"
  version: "1.0"
paths:
...
  /account/:
    get:
      tags:
        - account-controller
      operationId: getAllAccount
      description: "Retrieves all accounts"
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/AccountList"
    put:
      tags:
        - account-controller
      operationId: updateAccount
      description: "Updates an account"

      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
        required: true
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Account"
    post:
      tags:
        - account-controller
      operationId: createAccount
      description: "Creates an account"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
        required: true
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Account"
...
components:
  schemas:
    AccountList:
      type: array
      items:
        $ref: "#/components/schemas/Account"
      example:
        - id: 22214
          displayName: "test account 4"
        - id: 22215
          displayName: "test account 5"
...
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 22215
        displayName:
          type: string
          example: "test account 5"